refactor(cli,server): consolidate publish and stdlib-publish paths
Regular `tokf publish` did not inline `fixture = "file.txt"` references
in test files before uploading, unlike `tokf publish-stdlib`. This meant
regular publish could upload broken test files referencing fixtures the
server can't access. Both paths also duplicated filter validation logic
on the server side (`PreparedFilter` vs `PreparedStdlibFilter`).

CLI side:
- Extract `publish_shared.rs` with `hash_filter`, `inline_lua_script`,
  `resolve_fixtures_in_test`, and `collect_test_files_resolved`
- Both `publish_cmd` and `publish_stdlib_cmd` now use the shared helpers
- Regular publish now resolves fixture references (the key bug fix)

Server side:
- Extract `validate_and_prepare()` shared by both publish endpoints
- Delete `PreparedStdlibFilter`, reuse `PreparedFilter` with `pub(super)`
- `prepare_filter` and `prepare_stdlib_filter` delegate to shared fn
