+++ crates/tokf-cli/src/lib.rs
@@ -5,6 +5,7 @@ pub mod fs;
 pub mod history;
 pub mod hook;
 pub mod paths;
+pub mod publish_shared;
 pub mod remote;
 pub mod rewrite;
 pub mod runner;
