+++ crates/tokf-cli/src/publish_cmd.rs
@@ -1,8 +1,8 @@
 use std::io::BufRead as _;
-use std::path::Path;
 
 use tokf::auth::credentials;
 use tokf::config;
+use tokf::publish_shared::{collect_test_files_resolved, hash_filter, inline_lua_script};
 use tokf::remote::http::Client;
 use tokf::remote::publish_client;
 
@@ -43,76 +43,6 @@ fn resolve_local_filter(filter_name: &str) -> anyhow::Result<config::ResolvedFil
     Ok(resolved_filter)
 }
 
-/// Parse filter bytes and return `(content_hash, command_pattern)`.
-fn hash_filter(filter_bytes: &[u8]) -> anyhow::Result<(String, String)> {
-    let toml_str = std::str::from_utf8(filter_bytes)
-        .map_err(|_| anyhow::anyhow!("filter TOML is not valid UTF-8"))?;
-    let cfg: tokf_common::config::types::FilterConfig =
-        toml::from_str(toml_str).map_err(|e| anyhow::anyhow!("invalid filter TOML: {e}"))?;
-    let hash =
-        tokf_common::hash::canonical_hash(&cfg).map_err(|e| anyhow::anyhow!("hash error: {e}"))?;
-    Ok((hash, cfg.command.first().to_string()))
-}
-
-/// If `lua_script.file` is set, read the external file and embed its content
-/// as `lua_script.source` so the filter TOML is self-contained for publishing.
-///
-/// Returns the (possibly modified) filter bytes. When no `lua_script.file` is
-/// present, the original bytes are returned unchanged.
-///
-/// The script path is canonicalized and must reside within (or under) the
-/// filter file's parent directory to prevent path-traversal attacks.
-fn inline_lua_script(filter_bytes: Vec<u8>, filter_path: &Path) -> anyhow::Result<Vec<u8>> {
-    let toml_str = std::str::from_utf8(&filter_bytes)
-        .map_err(|_| anyhow::anyhow!("filter TOML is not valid UTF-8"))?;
-    let mut cfg: tokf_common::config::types::FilterConfig =
-        toml::from_str(toml_str).map_err(|e| anyhow::anyhow!("invalid filter TOML: {e}"))?;
-
-    let script_file = match cfg.lua_script.as_ref().and_then(|s| s.file.as_ref()) {
-        Some(f) => f.clone(),
-        None => return Ok(filter_bytes),
-    };
-
-    let base_dir = filter_path
-        .parent()
-        .unwrap_or_else(|| Path::new("."))
-        .canonicalize()
-        .map_err(|e| anyhow::anyhow!("cannot resolve filter directory: {e}"))?;
-    let script_path = base_dir.join(&script_file);
-    let canonical_script = script_path.canonicalize().map_err(|e| {
-        anyhow::anyhow!(
-            "cannot resolve lua_script.file '{}': {e}",
-            script_path.display()
-        )
-    })?;
-
-    if !canonical_script.starts_with(&base_dir) {
-        anyhow::bail!(
-            "lua_script.file '{}' escapes the filter directory — \
-             the script must reside within '{}'",
-            script_file,
-            base_dir.display()
-        );
-    }
-
-    let source = std::fs::read_to_string(&canonical_script).map_err(|e| {
-        anyhow::anyhow!(
-            "cannot read lua_script.file '{}': {e}",
-            canonical_script.display()
-        )
-    })?;
-
-    if let Some(script_cfg) = cfg.lua_script.as_mut() {
-        script_cfg.source = Some(source);
-        script_cfg.file = None;
-    }
-
-    eprintln!("[tokf] inlined lua_script.file '{script_file}' into filter source");
-    let serialized =
-        toml::to_string_pretty(&cfg).map_err(|e| anyhow::anyhow!("TOML serialize error: {e}"))?;
-    Ok(serialized.into_bytes())
-}
-
 // ── Publish flow ────────────────────────────────────────────────────────────
 
 fn publish(filter_name: &str, dry_run: bool) -> anyhow::Result<i32> {
@@ -122,7 +52,7 @@ fn publish(filter_name: &str, dry_run: bool) -> anyhow::Result<i32> {
     let filter_bytes = std::fs::read(&resolved_filter.source_path)?;
     let filter_bytes = inline_lua_script(filter_bytes, &resolved_filter.source_path)?;
     let (content_hash, command_pattern) = hash_filter(&filter_bytes)?;
-    let test_files = collect_test_files(&resolved_filter)?;
+    let test_files = collect_test_files_resolved(&resolved_filter.source_path)?;
 
     eprintln!("[tokf] publishing filter: {filter_name}");
     eprintln!("  Command: {command_pattern}");
@@ -160,7 +90,7 @@ fn publish_update_tests(filter_name: &str, dry_run: bool) -> anyhow::Result<i32>
 
     let filter_bytes = std::fs::read(&resolved_filter.source_path)?;
     let (content_hash, _) = hash_filter(&filter_bytes)?;
-    let test_files = collect_test_files(&resolved_filter)?;
+    let test_files = collect_test_files_resolved(&resolved_filter.source_path)?;
 
     if test_files.is_empty() {
         anyhow::bail!("no test files found for {filter_name}");
@@ -195,39 +125,6 @@ fn publish_update_tests(filter_name: &str, dry_run: bool) -> anyhow::Result<i32>
     );
     eprintln!("URL: {}", resp.registry_url);
     Ok(0)
-}
-
-fn collect_test_files(
-    resolved_filter: &config::ResolvedFilter,
-) -> anyhow::Result<Vec<(String, Vec<u8>)>> {
-    let stem = resolved_filter
-        .relative_path
-        .file_stem()
-        .unwrap_or_default()
-        .to_string_lossy();
-    let test_dir_name = format!("{stem}_test");
-    let source_test_dir = resolved_filter
-        .source_path
-        .parent()
-        .unwrap_or_else(|| Path::new("."))
-        .join(&test_dir_name);
-
-    if !source_test_dir.is_dir() {
-        return Ok(Vec::new());
-    }
-
-    let mut files = Vec::new();
-    for entry in std::fs::read_dir(&source_test_dir)? {
-        let entry = entry?;
-        let path = entry.path();
-        if path.is_file() {
-            let filename = entry.file_name().to_string_lossy().to_string();
-            let bytes = std::fs::read(&path)?;
-            files.push((filename, bytes));
-        }
-    }
-    files.sort_by(|a, b| a.0.cmp(&b.0));
-    Ok(files)
 }
 
 fn ensure_license_accepted() -> anyhow::Result<()> {
@@ -271,205 +168,5 @@ mod tests {
             .iter()
             .find(|f| f.matches_name("nonexistent/xyz-abc-filter-99"));
         assert!(found.is_none(), "expected no match for nonexistent filter");
-    }
-
-    #[test]
-    fn collect_test_files_from_adjacent_dir() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let filter_dir = dir.path().join("myns");
-        std::fs::create_dir_all(&filter_dir).unwrap();
-
-        // Create filter
-        let filter_path = filter_dir.join("my-filter.toml");
-        std::fs::write(&filter_path, r#"command = "my-cmd""#).unwrap();
-
-        // Create adjacent _test/ dir with test files
-        let test_dir = filter_dir.join("my-filter_test");
-        std::fs::create_dir_all(&test_dir).unwrap();
-        std::fs::write(test_dir.join("basic.toml"), b"name = \"basic\"").unwrap();
-        std::fs::write(test_dir.join("edge.toml"), b"name = \"edge\"").unwrap();
-
-        // Build a minimal ResolvedFilter pointing to our temp file
-        let cfg: tokf_common::config::types::FilterConfig =
-            toml::from_str(r#"command = "my-cmd""#).unwrap();
-        let hash = tokf_common::hash::canonical_hash(&cfg).unwrap_or_default();
-        let resolved = config::ResolvedFilter {
-            config: cfg,
-            hash,
-            source_path: filter_path,
-            relative_path: std::path::PathBuf::from("myns/my-filter.toml"),
-            priority: 0,
-        };
-
-        let files = collect_test_files(&resolved).unwrap();
-        assert_eq!(files.len(), 2, "expected 2 test files");
-        let names: std::collections::HashSet<_> = files.iter().map(|(n, _)| n.as_str()).collect();
-        assert!(names.contains("basic.toml"));
-        assert!(names.contains("edge.toml"));
-    }
-
-    #[test]
-    fn inline_lua_script_embeds_file_content() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let filter_path = dir.path().join("my-filter.toml");
-        let script_path = dir.path().join("transform.luau");
-
-        std::fs::write(&script_path, "return output:upper()").unwrap();
-        std::fs::write(
-            &filter_path,
-            r#"command = "my-cmd"
-
-[lua_script]
-lang = "luau"
-file = "transform.luau"
-"#,
-        )
-        .unwrap();
-
-        let filter_bytes = std::fs::read(&filter_path).unwrap();
-        let result = inline_lua_script(filter_bytes, &filter_path).unwrap();
-
-        let cfg: tokf_common::config::types::FilterConfig =
-            toml::from_str(std::str::from_utf8(&result).unwrap()).unwrap();
-        let script = cfg.lua_script.unwrap();
-        assert!(script.file.is_none(), "file should be removed");
-        assert_eq!(script.source.unwrap(), "return output:upper()");
-    }
-
-    #[test]
-    fn inline_lua_script_rejects_path_traversal() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let subdir = dir.path().join("filters");
-        std::fs::create_dir_all(&subdir).unwrap();
-
-        // Create a file outside the filter directory
-        let secret = dir.path().join("secret.txt");
-        std::fs::write(&secret, "sensitive data").unwrap();
-
-        let filter_path = subdir.join("my-filter.toml");
-        std::fs::write(
-            &filter_path,
-            r#"command = "my-cmd"
-
-[lua_script]
-lang = "luau"
-file = "../secret.txt"
-"#,
-        )
-        .unwrap();
-
-        let filter_bytes = std::fs::read(&filter_path).unwrap();
-        let result = inline_lua_script(filter_bytes, &filter_path);
-        assert!(result.is_err(), "path traversal should be rejected");
-        let msg = result.unwrap_err().to_string();
-        assert!(
-            msg.contains("escapes the filter directory"),
-            "expected traversal error, got: {msg}"
-        );
-    }
-
-    #[test]
-    fn inline_lua_script_allows_subdirectory() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let scripts_dir = dir.path().join("scripts");
-        std::fs::create_dir_all(&scripts_dir).unwrap();
-
-        std::fs::write(scripts_dir.join("helper.luau"), "return 'ok'").unwrap();
-
-        let filter_path = dir.path().join("my-filter.toml");
-        std::fs::write(
-            &filter_path,
-            r#"command = "my-cmd"
-
-[lua_script]
-lang = "luau"
-file = "scripts/helper.luau"
-"#,
-        )
-        .unwrap();
-
-        let filter_bytes = std::fs::read(&filter_path).unwrap();
-        let result = inline_lua_script(filter_bytes, &filter_path).unwrap();
-
-        let cfg: tokf_common::config::types::FilterConfig =
-            toml::from_str(std::str::from_utf8(&result).unwrap()).unwrap();
-        let script = cfg.lua_script.unwrap();
-        assert!(script.file.is_none());
-        assert_eq!(script.source.unwrap(), "return 'ok'");
-    }
-
-    #[test]
-    fn inline_lua_script_empty_file_produces_empty_source() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let filter_path = dir.path().join("my-filter.toml");
-        std::fs::write(dir.path().join("empty.luau"), "").unwrap();
-        std::fs::write(
-            &filter_path,
-            r#"command = "my-cmd"
-
-[lua_script]
-lang = "luau"
-file = "empty.luau"
-"#,
-        )
-        .unwrap();
-
-        let filter_bytes = std::fs::read(&filter_path).unwrap();
-        let result = inline_lua_script(filter_bytes, &filter_path).unwrap();
-
-        let cfg: tokf_common::config::types::FilterConfig =
-            toml::from_str(std::str::from_utf8(&result).unwrap()).unwrap();
-        let script = cfg.lua_script.unwrap();
-        assert_eq!(script.source.unwrap(), "");
-    }
-
-    #[test]
-    fn inline_lua_script_hash_stable_when_no_file() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let filter_path = dir.path().join("my-filter.toml");
-        let toml_str = r#"command = "my-cmd""#;
-        std::fs::write(&filter_path, toml_str).unwrap();
-
-        let filter_bytes = toml_str.as_bytes().to_vec();
-        let (hash_before, _) = hash_filter(&filter_bytes).unwrap();
-        let result = inline_lua_script(filter_bytes, &filter_path).unwrap();
-        let (hash_after, _) = hash_filter(&result).unwrap();
-        assert_eq!(
-            hash_before, hash_after,
-            "hash should be stable when no inlining occurs"
-        );
-    }
-
-    #[test]
-    fn inline_lua_script_noop_without_file() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let filter_path = dir.path().join("my-filter.toml");
-        let toml_str = r#"command = "my-cmd""#;
-        std::fs::write(&filter_path, toml_str).unwrap();
-
-        let filter_bytes = toml_str.as_bytes().to_vec();
-        let result = inline_lua_script(filter_bytes.clone(), &filter_path).unwrap();
-        assert_eq!(result, filter_bytes, "should return unchanged bytes");
-    }
-
-    #[test]
-    fn collect_test_files_returns_empty_when_no_test_dir() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let filter_path = dir.path().join("my-filter.toml");
-        std::fs::write(&filter_path, r#"command = "my-cmd""#).unwrap();
-
-        let cfg: tokf_common::config::types::FilterConfig =
-            toml::from_str(r#"command = "my-cmd""#).unwrap();
-        let hash = tokf_common::hash::canonical_hash(&cfg).unwrap_or_default();
-        let resolved = config::ResolvedFilter {
-            config: cfg,
-            hash,
-            source_path: filter_path,
-            relative_path: std::path::PathBuf::from("my-filter.toml"),
-            priority: 0,
-        };
-
-        let files = collect_test_files(&resolved).unwrap();
-        assert!(files.is_empty());
     }
 }
