+++ crates/tokf-cli/src/publish_stdlib_cmd.rs
@@ -2,6 +2,7 @@ use std::path::{Path, PathBuf};
 
 use serde::{Deserialize, Serialize};
 
+use tokf::publish_shared::collect_test_files_resolved;
 use tokf::remote::http::Client;
 
 /// Entry point for the `tokf publish-stdlib` subcommand.
@@ -161,13 +162,21 @@ fn collect_stdlib_entries(filters_dir: &Path) -> anyhow::Result<Vec<StdlibFilter
     for path in &filter_paths {
         let filter_toml = std::fs::read_to_string(path)?;
         let author = resolve_author(path).unwrap_or_else(|| fallback_author.clone());
-        let test_files = collect_and_resolve_tests(path)?;
+
+        let raw_tests = collect_test_files_resolved(path)?;
+        let test_files = raw_tests
+            .into_iter()
+            .map(|(filename, bytes)| StdlibTestFile {
+                filename,
+                content: String::from_utf8_lossy(&bytes).to_string(),
+            })
+            .collect();
 
         eprintln!(
             "[publish-stdlib]   {} (author: {}, tests: {})",
             path.display(),
             author,
-            test_files.len()
+            count_test_files(&test_files),
         );
 
         entries.push(StdlibFilterEntry {
@@ -178,6 +187,10 @@ fn collect_stdlib_entries(filters_dir: &Path) -> anyhow::Result<Vec<StdlibFilter
     }
 
     Ok(entries)
+}
+
+fn count_test_files(files: &[StdlibTestFile]) -> usize {
+    files.len()
 }
 
 /// Recursively find all *.toml files that aren't inside _test/ directories.
@@ -195,96 +208,6 @@ fn collect_filter_files(dir: &Path, out: &mut Vec<PathBuf>) -> anyhow::Result<()
         }
     }
     Ok(())
-}
-
-/// Collect test TOML files from the adjacent `{stem}_test/` directory,
-/// resolving `fixture = "file.txt"` references to `inline` content.
-fn collect_and_resolve_tests(filter_path: &Path) -> anyhow::Result<Vec<StdlibTestFile>> {
-    let stem = filter_path
-        .file_stem()
-        .unwrap_or_default()
-        .to_string_lossy();
-    let test_dir = filter_path
-        .parent()
-        .unwrap_or_else(|| Path::new("."))
-        .join(format!("{stem}_test"));
-
-    if !test_dir.is_dir() {
-        return Ok(Vec::new());
-    }
-
-    let mut files = Vec::new();
-    for entry in std::fs::read_dir(&test_dir)? {
-        let entry = entry?;
-        let path = entry.path();
-        if !path.is_file() {
-            continue;
-        }
-        let filename = entry.file_name().to_string_lossy().to_string();
-        if path.extension().is_none_or(|e| e != "toml") {
-            continue;
-        }
-
-        let content = std::fs::read_to_string(&path)?;
-        let resolved = resolve_fixtures_in_test(&content, &test_dir)
-            .map_err(|e| anyhow::anyhow!("{}: {e}", path.display()))?;
-
-        files.push(StdlibTestFile {
-            filename,
-            content: resolved,
-        });
-    }
-    files.sort_by(|a, b| a.filename.cmp(&b.filename));
-    Ok(files)
-}
-
-/// Replace `fixture = "file.txt"` with `inline = "<file contents>"` in a test
-/// TOML string. Fixture files are resolved relative to `test_dir`.
-fn resolve_fixtures_in_test(content: &str, test_dir: &Path) -> anyhow::Result<String> {
-    // Parse to check if there's a fixture reference
-    let case: tokf_common::test_case::TestCase =
-        toml::from_str(content).map_err(|e| anyhow::anyhow!("invalid test TOML: {e}"))?;
-
-    let Some(fixture_name) = case.fixture else {
-        // No fixture — return as-is
-        return Ok(content.to_string());
-    };
-
-    if case.inline.is_some() {
-        // Has both — inline takes priority, return as-is
-        return Ok(content.to_string());
-    }
-
-    let fixture_path = test_dir.join(&fixture_name);
-    let fixture_content = std::fs::read_to_string(&fixture_path)
-        .map_err(|e| anyhow::anyhow!("cannot read fixture '{fixture_name}': {e}"))?;
-    let fixture_content = fixture_content.trim_end();
-
-    // Rebuild the TOML with inline instead of fixture
-    let mut result = String::new();
-    let mut skipped_fixture = false;
-    for line in content.lines() {
-        let trimmed = line.trim_start();
-        if trimmed.starts_with("fixture") && trimmed.contains('=') {
-            skipped_fixture = true;
-            continue;
-        }
-        result.push_str(line);
-        result.push('\n');
-    }
-
-    if !skipped_fixture {
-        anyhow::bail!("expected fixture line but did not find it");
-    }
-
-    // Insert inline using TOML multi-line literal string (triple single-quotes).
-    // Literal strings don't interpret backslashes, which is critical for fixture
-    // content that may contain paths, regex patterns, or separator lines like \---.
-    let insert_pos = result.find("[[expect]]").unwrap_or(result.len());
-    let inline_value = format!("inline = '''\n{fixture_content}\n'''\n");
-    result.insert_str(insert_pos, &inline_value);
-
-    Ok(result)
 }
 
 /// Best-effort: resolve the GitHub username of the last committer for a file.
@@ -333,89 +256,4 @@ fn resolve_fallback_author() -> String {
         }
     }
     "mpecan".to_string()
-}
-
-#[cfg(test)]
-#[allow(clippy::unwrap_used)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn resolve_fixtures_replaces_fixture_with_inline() {
-        let dir = tempfile::TempDir::new().unwrap();
-        std::fs::write(dir.path().join("output.txt"), "hello world\n").unwrap();
-
-        let toml = r#"name = "test"
-fixture = "output.txt"
-exit_code = 0
-
-[[expect]]
-contains = "hello"
-"#;
-        let resolved = resolve_fixtures_in_test(toml, dir.path()).unwrap();
-        assert!(!resolved.contains("fixture"), "fixture should be removed");
-        assert!(resolved.contains("inline"), "should contain inline");
-        assert!(
-            resolved.contains("hello world"),
-            "should contain fixture content"
-        );
-    }
-
-    #[test]
-    fn resolve_fixtures_preserves_inline() {
-        let toml = r#"name = "test"
-inline = "already inline"
-
-[[expect]]
-equals = "already inline"
-"#;
-        let resolved = resolve_fixtures_in_test(toml, Path::new("/nonexistent")).unwrap();
-        assert_eq!(resolved, toml, "should be unchanged");
-    }
-
-    #[test]
-    fn resolve_fixtures_uses_literal_strings_for_backslashes() {
-        let dir = tempfile::TempDir::new().unwrap();
-        // Fixture content with backslashes (like gradle's \--- separators)
-        std::fs::write(dir.path().join("output.txt"), "\\--- foo\n\\--- bar\n").unwrap();
-
-        let toml = r#"name = "test"
-fixture = "output.txt"
-exit_code = 0
-
-[[expect]]
-contains = "foo"
-"#;
-        let resolved = resolve_fixtures_in_test(toml, dir.path()).unwrap();
-        // Must use ''' (literal string) so backslashes aren't interpreted
-        assert!(
-            resolved.contains("'''"),
-            "should use literal string quotes, got:\n{resolved}"
-        );
-        assert!(
-            !resolved.contains(r#"""""#),
-            "should NOT use basic string quotes"
-        );
-        assert!(
-            resolved.contains("\\--- foo"),
-            "backslashes should be preserved literally"
-        );
-    }
-
-    #[test]
-    fn resolve_fixtures_errors_on_missing_file() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let toml = r#"name = "test"
-fixture = "missing.txt"
-
-[[expect]]
-contains = "x"
-"#;
-        let result = resolve_fixtures_in_test(toml, dir.path());
-        assert!(result.is_err());
-        assert!(
-            result.unwrap_err().to_string().contains("missing.txt"),
-            "error should mention the file"
-        );
-    }
 }
